# api-spec.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Passport Agent — Agent Control Plane API",
    "version": "2.1.0",
    "description": "AI agent identity + permission enforcement API. 15 endpoints for agent lifecycle management, task execution, policy enforcement, and observability."
  },
  "servers": [{ "url": "http://localhost:3000", "description": "Local development" }],
  "paths": {
...
    "/agents/register": {
      "post": { "summary": "Register a new agent", "security": [{ "bearerAuth": [] }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "name": { "type": "string" }, "model": { "type": "string" }, "provider": { "type": "string" }, "systemPrompt": { "type": "string" } }, "required": ["name","model","provider"] } } } }, "responses": { "201": { "description": "Agent registered with secret key" } } }
    },
...
    "/agents/{id}": {
      "get": { "summary": "Get agent by ID", "parameters": [{ "name": "id", "in": "path", "required": true }], "responses": { "200": { "description": "Agent document" }, "404": { "description": "Not found" } } },
      "patch": { "summary": "Revoke agent", "security": [{ "bearerAuth": [] }], "parameters": [{ "name": "id", "in": "path", "required": true }], "responses": { "200": { "description": "Agent revoked" } } }
    },
...
    "/enforce": {
      "post": { "summary": "Evaluate agent intent against policies", "security": [{ "bearerAuth": [] }], "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "intent": { "type": "object", "properties": { "intentId": { "type": "string" }, "agentId": { "type": "string" }, "tool": { "type": "string" }, "parameters": { "type": "object" } }, "required": ["intentId","agentId","tool"] } } } } } }, "responses": { "200": { "description": "Enforcement decision + gateway ticket" } } }
    },
    "/gateway/execute": {
      "post": { "summary": "Execute tool via gateway ticket", "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "gatewayTicket": { "type": "string" }, "action": { "type": "object" } }, "required": ["gatewayTicket","action"] } } } }, "responses": { "200": { "description": "Executed" }, "403": { "description": "Invalid/replayed ticket" } } }
    },
...
    "/run/{id}/trace": {
      "get": { "summary": "Full execution trace for a run", "parameters": [{ "name": "id", "in": "path", "required": true }], "responses": { "200": { "description": "Run + task + events" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  }
}